+++ repose-aggregator/tests/functional-tests/src/integrationTest/groovy/features/filters/keystonev2/akkatimeout/AkkaTimeoutSameAsHttpConnTimeoutTest.groovy
@@ -111,7 +111,7 @@ class AkkaTimeoutSameAsHttpConnTimeoutTest extends ReposeValveTest {
         )
 
         then: "Request should not be passed from repose"
-        mc.receivedResponse.code == HttpServletResponse.SC_GATEWAY_TIMEOUT
+        mc.receivedResponse.code as Integer == HttpServletResponse.SC_GATEWAY_TIMEOUT
         mc.handlings.size() == 0
         sleep(1000)
         reposeLogSearch.searchByString("Error acquiring value from akka .* or the cache. Reason: Futures timed out after .31000 milliseconds.").size() > 0
